Clean up weird mac API
